pyproject.toml
[project]
name = "thesis-format-engine"
version = "0.1.0"
description = "Rule-driven thesis formatting inspection and repair engine for DOCX documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "lxml>=5.3.0",
  "pydantic>=2.10.0",
  "PyYAML>=6.0.2",
  "python-docx>=1.1.2",
  "typer>=0.15.1",
  "rich>=13.9.4"
]

[project.scripts]
thesisfix = "thesis_format_engine.cli.main:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
